Refactor TimeTable Unit property handling
Refactored the `Unit` property in the `TimeTable` class to use a backing field with a default value of `"µs"`. Updated the `get` accessor to return the backing field and the `set` accessor to update the `"Time"` column's `HeaderText` based on the value. Removed the previous implementation that directly manipulated the `HeaderText` for extracting and setting the unit.