Improve OrderedSortedDictionary key handling logic
Updated the `set` accessor of the `this[TKey]` indexer to allow dynamic insertion of new key-value pairs when a key does not exist, replacing the previous behavior that threw a `KeyNotFoundException`. If the key exists, the value is updated as before. This change enhances the flexibility and usability of the `OrderedSortedDictionary<TKey, TValue>` class.